Little bit of comms rework to support sending buffers instead of individual bytes. Will be useful for i2c

+++ c/projects/timing_block/include/usb_comms.h
@@ -11,7 +11,7 @@
 #include "io/comms.h"
 
 void usb_comms_init(void);
-bool usb_comms_write(uint8_t *msg, uint32_t datalen);
+bool usb_comms_write(uint8_t *msg, uint16_t datalen);
 bool usb_comms_write_byte(uint8_t data);
 void usb_comms_publish_blocking(comms_channel_t channel, uint8_t *msg, uint16_t msg_len);
 void usb_comms_subscribe(comms_channel_t channel, subscriber_t subscriber, void *usr);
